new: refactor async client generator
