+++ tools/async_client_generator/config.py
@@ -1,4 +1,5 @@
 from pathlib import Path
 
-ROOT_DIR = Path(__file__).parent.parent.parent
-CODE_DIR = ROOT_DIR / "qdrant_client"
+CODE_DIR = Path(__file__).parent
+ROOT_DIR = CODE_DIR.parent.parent
+CLIENT_DIR = ROOT_DIR / "qdrant_client"
