+++ tools/async_client_generator/transformers/fastembed/function_def_transformer.py
@@ -1,0 +1,6 @@
+from ...transformers import FunctionDefTransformer
+
+
+class FastembedFunctionDefTransformer(FunctionDefTransformer):
+    def _keep_sync(self, name: str) -> bool:
+        return name in self.keep_sync or name.startswith("_")
